# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Document Automation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Employee {
  id: number;
  fullName: string;
  position: string;
  department: string;
  employeeNumber: string;
  hireDate: string;
  salary?: number;
  phone?: string;
  email?: string;
  createdAt?: string;
}

export interface CreateEmployee {
  fullName: string;
  position: string;
  department: string;
  employeeNumber: string;
  hireDate: string;
  salary?: number;
  phone?: string;
  email?: string;
}

export type DocumentStatus =
  (typeof DocumentStatus)[keyof typeof DocumentStatus];

export const DocumentStatus = {
  draft: "draft",
  signed: "signed",
  printed: "printed",
} as const;

export interface Document {
  id: number;
  number: string;
  type: string;
  title: string;
  status: DocumentStatus;
  employeeId: number;
  employeeName: string;
  content: string;
  createdAt: string;
  signedAt?: string;
  templateId?: number;
}

export type CreateDocumentFields = { [key: string]: string };

export interface CreateDocument {
  templateId: number;
  employeeId: number;
  fields: CreateDocumentFields;
}

export type TemplateFieldType =
  (typeof TemplateFieldType)[keyof typeof TemplateFieldType];

export const TemplateFieldType = {
  text: "text",
  date: "date",
  number: "number",
  select: "select",
} as const;

export interface TemplateField {
  key: string;
  label: string;
  type: TemplateFieldType;
  required: boolean;
  options?: string[];
}

export interface Template {
  id: number;
  name: string;
  type: string;
  description: string;
  fields: TemplateField[];
  content: string;
}

export interface Stats {
  totalDocuments: number;
  draftDocuments: number;
  signedDocuments: number;
  printedDocuments: number;
  totalEmployees: number;
  documentsThisMonth: number;
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}

export type GetDocumentsParams = {
  type?: string;
  status?: string;
  employeeId?: number;
};