# lib/api-zod/src/generated/types/createDocument.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Document Automation API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateDocumentFields } from "./createDocumentFields";

export interface CreateDocument {
  templateId: number;
  employeeId: number;
  fields: CreateDocumentFields;
}