# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Document Automation API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentStatus } from "./documentStatus";

export interface Document {
  id: number;
  number: string;
  type: string;
  title: string;
  status: DocumentStatus;
  employeeId: number;
  employeeName: string;
  content: string;
  createdAt: string;
  signedAt?: string;
  templateId?: number;
}