# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Document Automation API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  fullName: string;
  position: string;
  department: string;
  employeeNumber: string;
  hireDate: string;
  salary?: number;
  phone?: string;
  email?: string;
  createdAt?: string;
}