# lib/api-zod/src/generated/types/template.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Document Automation API
 * OpenAPI spec version: 0.1.0
 */
import type { TemplateField } from "./templateField";

export interface Template {
  id: number;
  name: string;
  type: string;
  description: string;
  fields: TemplateField[];
  content: string;
}